% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_distance}
\alias{compute_distance}
\title{compute_distance}
\usage{
compute_distance(dat, type, totcounts)
}
\arguments{
\item{dat}{}

\item{type}{}

\item{totcounts}{}
}
\description{
compute_distance
}
