% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_probability_mass}
\alias{generate_probability_mass}
\title{generate_probability_mass}
\usage{
generate_probability_mass(n_subtypes, frac_hk_genes, fold_changes)
}
\arguments{
\item{n_subtypes}{}

\item{frac_hk_genes}{}

\item{fold_changes}{}
}
\description{
generate_probability_mass
}
