% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{log_likelihood_ratio_multinomial}
\alias{log_likelihood_ratio_multinomial}
\title{log_likelihood_ratio_multinomial}
\usage{
log_likelihood_ratio_multinomial(prob1, prob2, totcounts)
}
\arguments{
\item{prob1}{}

\item{prob2}{}

\item{totcounts}{}
}
\description{
log_likelihood_ratio_multinomial
}
