% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sequence_cell}
\alias{sequence_cell}
\title{sequence_cell}
\usage{
sequence_cell(ncells, reads_per_cell, prob_dist, seed)
}
\arguments{
\item{ncells}{number of cells to generate count vectors for}

\item{reads_per_cell}{number of reads/UMIs per cell (assumed equal for all cells)}

\item{prob_dist}{probability mass (relative abundance of genes)}

\item{seed}{random seed}
}
\value{
The object returned is the MLE of probability mass 
(restrict to this for this study, as do not require actual 
count vectors for most analysis)
}
\description{
Generate count vector for a number of cells of a 
fixed subpopulation, given fixed number of reads and a probability mass.
}
